backend/requirements.txt
# --- CORE FRAMEWORK ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0

# --- NETWORKING & SCRAPING ---
httpx>=0.27.0
beautifulsoup4>=4.12.3
requests>=2.31.0
aiohttp>=3.9.3

# --- SECURITY & AUTH ---
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# --- UTILITIES ---
python-dotenv>=1.0.1
loguru>=0.7.2
websockets>=12.0

# --- OPTIONAL ---
redis>=5.0.3